The amigaBlock class

The Commodore Amiga stores data on floppy disks as 512 byte blocks. This class reflects such a block.

Details

There are several types of blocks. Most important are the boot block (used for booting the Amiga system), the root block (containing information on the disk and the root directory), header blocks (indicating where to find file data) and data blocks (containing the actual file data). See this package's vignette (vignette("amigaDiskFiles")) for more details. use the amigaBlock-method to extract a specific block from an amigaDisk object.

Slots

data

The raw data of a 'block' of data on an Amiga disk file. Each block holds 512 bytes of information. This slot is therefore a vector of the same length.

Author

Pepijn de Vries

Examples

## create a block with no data:
new("amigaBlock")
#> 0x000  00000000 00000000 00000000 00000000  ................
#> 0x010  00000000 00000000 00000000 00000000  ................
#> 0x020  00000000 00000000 00000000 00000000  ................
#> 0x030  00000000 00000000 00000000 00000000  ................
#> 0x040  00000000 00000000 00000000 00000000  ................
#> 0x050  00000000 00000000 00000000 00000000  ................
#> 0x060  00000000 00000000 00000000 00000000  ................
#> 0x070  00000000 00000000 00000000 00000000  ................
#> 0x080  00000000 00000000 00000000 00000000  ................
#> 0x090  00000000 00000000 00000000 00000000  ................
#> 0x0A0  00000000 00000000 00000000 00000000  ................
#> 0x0B0  00000000 00000000 00000000 00000000  ................
#> 0x0C0  00000000 00000000 00000000 00000000  ................
#> 0x0D0  00000000 00000000 00000000 00000000  ................
#> 0x0E0  00000000 00000000 00000000 00000000  ................
#> 0x0F0  00000000 00000000 00000000 00000000  ................
#> 0x100  00000000 00000000 00000000 00000000  ................
#> 0x110  00000000 00000000 00000000 00000000  ................
#> 0x120  00000000 00000000 00000000 00000000  ................
#> 0x130  00000000 00000000 00000000 00000000  ................
#> 0x140  00000000 00000000 00000000 00000000  ................
#> 0x150  00000000 00000000 00000000 00000000  ................
#> 0x160  00000000 00000000 00000000 00000000  ................
#> 0x170  00000000 00000000 00000000 00000000  ................
#> 0x180  00000000 00000000 00000000 00000000  ................
#> 0x190  00000000 00000000 00000000 00000000  ................
#> 0x1A0  00000000 00000000 00000000 00000000  ................
#> 0x1B0  00000000 00000000 00000000 00000000  ................
#> 0x1C0  00000000 00000000 00000000 00000000  ................
#> 0x1D0  00000000 00000000 00000000 00000000  ................
#> 0x1E0  00000000 00000000 00000000 00000000  ................
#> 0x1F0  00000000 00000000 00000000 00000000  ................