Get the block ID from the physical location on the disk

Get the block identifier based on the physical location on a disk (side, cylinder and sector) and the disk type.

Usage

# S4 method for class 'character,numeric,numeric,numeric'
get.blockID(disktype, sector, side, cylinder)

Arguments

disktype

A character string indicating the type of disk: DD for double density disks. HD for high density disks.

sector

numeric identifier for the sector on the disk, ranging from 0 up to 10 (DD disks) or 21 (HD disks).

side

numeric identifier for the side of the disk (0 or 1).

cylinder

numeric identifier for the cylinder on the disk, ranging from 0 up to 79.

Value

Returns the numeric identifier for the corresponding block.

Details

Data on Amiga floppy disks are stored as 512 byte blocks. These blocks are physically stored on a specific cylinder and side at a specific sector. This method returns the block identifier based on the physical location on the disk. The inverse of this function is achieved with the get.diskLocation method.

Note that all identifiers (or indices) have a base at zero, for consistency with Amiga specifications and documentation, opposed to the base of one used in R.

See also

Author

Pepijn de Vries

Examples

## Get the block identifier for sectors 0 up to 3 combined with
## cylinders 0 up to 3 on side 0 of the disk:
get.blockID(disktype = "DD",
            sector   = 0:3,
            side     = 0,
            cylinder = 0:3)
#> [1]  0 23 46 69