Convert raw values into Amiga integers

Convert raw data into 8, 16, or 32-bit signed or unsigned integer values, conform Amiga specifications.

Usage

rawToAmigaInt(x, bits = 8, signed = F)

Arguments

x

A vector of class raw to be converted into a character.

bits

Number of bits that represents the integer value. Should be 8 or a positive multitude of 8.

signed

A logical value indicating whether the integer should be signed (TRUE, default) or not (FALSE).

Value

A numeric value (or a vector of values), representing the integer data represented by the provided raw data. Note that R defines integer as 32-bit signed integers and cannot store the 32-bit signed values. Therefore a numeric value is returned rather than an explicit integer.

Details

The Commodore Amiga has specified the following data formats to represent integer data: BYTE (signed 8-bit integer), UBYTE (unsigned 8-bit integer), WORD (signed 16-bit integer), UWORD (unsigned 16-bit integer), LONG (signed 32-bit integer), ULONG, (unsigned 32-bit integer). This function converts raw data into such integers. Note that WORD and UWORD are also referred to as SHORT and USHORT respectively.

See also

Author

Pepijn de Vries

Examples

## Let's start by obtaining unsigned 8-bit integers:
rawToAmigaInt(as.raw(0:255))
#>   [1]   0   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#>  [19]  18  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35
#>  [37]  36  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53
#>  [55]  54  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71
#>  [73]  72  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89
#>  [91]  90  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107
#> [109] 108 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125
#> [127] 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143
#> [145] 144 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161
#> [163] 162 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179
#> [181] 180 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197
#> [199] 198 199 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215
#> [217] 216 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233
#> [235] 234 235 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251
#> [253] 252 253 254 255

## note that this is the same as:
as.numeric(as.raw(0:255))
#>   [1]   0   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#>  [19]  18  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35
#>  [37]  36  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53
#>  [55]  54  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71
#>  [73]  72  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89
#>  [91]  90  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107
#> [109] 108 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125
#> [127] 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143
#> [145] 144 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161
#> [163] 162 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179
#> [181] 180 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197
#> [199] 198 199 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215
#> [217] 216 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233
#> [235] 234 235 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251
#> [253] 252 253 254 255

## but with this function we can also get signed values:
rawToAmigaInt(as.raw(0:255), signed = TRUE)
#>   [1]    0    1    2    3    4    5    6    7    8    9   10   11   12   13   14
#>  [16]   15   16   17   18   19   20   21   22   23   24   25   26   27   28   29
#>  [31]   30   31   32   33   34   35   36   37   38   39   40   41   42   43   44
#>  [46]   45   46   47   48   49   50   51   52   53   54   55   56   57   58   59
#>  [61]   60   61   62   63   64   65   66   67   68   69   70   71   72   73   74
#>  [76]   75   76   77   78   79   80   81   82   83   84   85   86   87   88   89
#>  [91]   90   91   92   93   94   95   96   97   98   99  100  101  102  103  104
#> [106]  105  106  107  108  109  110  111  112  113  114  115  116  117  118  119
#> [121]  120  121  122  123  124  125  126  127 -128 -127 -126 -125 -124 -123 -122
#> [136] -121 -120 -119 -118 -117 -116 -115 -114 -113 -112 -111 -110 -109 -108 -107
#> [151] -106 -105 -104 -103 -102 -101 -100  -99  -98  -97  -96  -95  -94  -93  -92
#> [166]  -91  -90  -89  -88  -87  -86  -85  -84  -83  -82  -81  -80  -79  -78  -77
#> [181]  -76  -75  -74  -73  -72  -71  -70  -69  -68  -67  -66  -65  -64  -63  -62
#> [196]  -61  -60  -59  -58  -57  -56  -55  -54  -53  -52  -51  -50  -49  -48  -47
#> [211]  -46  -45  -44  -43  -42  -41  -40  -39  -38  -37  -36  -35  -34  -33  -32
#> [226]  -31  -30  -29  -28  -27  -26  -25  -24  -23  -22  -21  -20  -19  -18  -17
#> [241]  -16  -15  -14  -13  -12  -11  -10   -9   -8   -7   -6   -5   -4   -3   -2
#> [256]   -1

## Furthermore 32 or 16-bit integers can also be obtained.
## Let's look at 16-bit integers:
rawToAmigaInt(as.raw(0:255), 16)
#>   [1]     1   515  1029  1543  2057  2571  3085  3599  4113  4627  5141  5655
#>  [13]  6169  6683  7197  7711  8225  8739  9253  9767 10281 10795 11309 11823
#>  [25] 12337 12851 13365 13879 14393 14907 15421 15935 16449 16963 17477 17991
#>  [37] 18505 19019 19533 20047 20561 21075 21589 22103 22617 23131 23645 24159
#>  [49] 24673 25187 25701 26215 26729 27243 27757 28271 28785 29299 29813 30327
#>  [61] 30841 31355 31869 32383 32897 33411 33925 34439 34953 35467 35981 36495
#>  [73] 37009 37523 38037 38551 39065 39579 40093 40607 41121 41635 42149 42663
#>  [85] 43177 43691 44205 44719 45233 45747 46261 46775 47289 47803 48317 48831
#>  [97] 49345 49859 50373 50887 51401 51915 52429 52943 53457 53971 54485 54999
#> [109] 55513 56027 56541 57055 57569 58083 58597 59111 59625 60139 60653 61167
#> [121] 61681 62195 62709 63223 63737 64251 64765 65279

## Note that 16-bit integers require twice as many bytes
## as 8 bit integers:
length(rawToAmigaInt(as.raw(0:255), 16))
#> [1] 128
length(rawToAmigaInt(as.raw(0:255), 8))
#> [1] 256