The amigaDisk class

An S4 class representing the information from an Amiga Disk File.

Details

An Amiga Disk File (ADF) holds the raw data of an Amiga disk in the same order as blocks (amigaBlock) on the physical disks. As an Amiga Disk can hold any kind of information, so can this class.

An ADF file does not hold any other information. The size of the file will dictate whether it represents a double density floppy disk (880 kB) or a high density floppy disk (1760 kB). The disk type is also stored in this class.

Finally, the current directory is stored with this class. Which is only useful for DOS-formatted disks (with a file structure). By default this is set to the disk's root.

For more (technical) backgrounds please check this package's vignette (vignette("amigaDiskFiles"))

Use the objects constructor (new("amigaDisk")) to create a completely blank disk (without a filesystem). If you want to be able to transfer files from and to the virtual disk, use blank.amigaDOSDisk instead.

Slots

data

The raw data of the virtual disk. It should be a vector of length 901,120 in case of a double density disk and 1,802,240 in case of a high density disk.

type

A character indicating whether the virtual disk represents a "DD" (double density, most common) or "HD" (high density) disk.

current.dir

An integer, pointing at the block address of the current directory of this virtual disk. Use current.adf.dir to get or set the current directory.

Author

Pepijn de Vries

Examples

## This creates a blank non-bootable, non-DOS disk:
new("amigaDisk")
#> 
#> Amiga (DD) Disk File:
#> 	Type:			Non-bootable Non-DOS